requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
tensorflow==2.16.1
pillow==10.3.0
numpy==1.26.4
requests==2.31.0
gunicorn==21.2.0
